Move client/server logic to dedicated classes, Add REPL commands input
